refactor(api): move complementary certification habilitation repository
in src/organizational-entities
