+++ api/src/organizational-entities/infrastructure/repositories/complementary-certification-habilitation.repository.js
@@ -1,4 +1,4 @@
-import { DomainTransaction } from '../DomainTransaction.js';
+import { DomainTransaction } from '../../../shared/domain/DomainTransaction.js';
 
 const COMPLEMENTARY_CERTIFICATION_HABILITATIONS_TABLE_NAME = 'complementary-certification-habilitations';
